contrib/log/slog: avoid logging 0 value trace/span IDs
If the tracer is started but explicitly deactivated, there will be a
span on the context, but the span ID and trace ID will be set to 0.
Adding this information to log outputs seems like unintentional noise
(although it's certainly possible there's a use case I'm not aware of),
so I figured disabling it would be preferred.

The change to explicitly add `tracer.WithTraceEnabled(true)` was so that
the test would be resilient to folks who might have `DD_TRACE_ENABLED`
set explicitly, which is true on my machine.